+++ example/macos/Runner/MainFlutterWindow.swift
@@ -1,5 +1,7 @@
 import Cocoa
 import FlutterMacOS
+// N'oubliez pas d'importer le module de votre plugin si nécessaire
+import multi_screen_presentation
 
 class MainFlutterWindow: NSWindow {
   override func awakeFromNib() {
@@ -8,7 +10,13 @@ class MainFlutterWindow: NSWindow {
     self.contentViewController = flutterViewController
     self.setFrame(windowFrame, display: true)
 
+    // Enregistrement sur la vue principale
     RegisterGeneratedPlugins(registry: flutterViewController)
+
+    // Étape 3 : Transmettre la référence de la fonction générée au plugin
+    MultiScreenPresentationPlugin.registerPluginsHook = { registry in
+      RegisterGeneratedPlugins(registry: registry)
+    }
 
     super.awakeFromNib()
   }